refactor: move OTW firmware updates to Driver class
