+++ docs/getting-started/migrating/v15.md
@@ -25,6 +25,21 @@ As of this release, Z-Wave JS supports communicating with the text-based CLI exp
 - The `driver.isInBootloader()` method has been removed. Use the new `driver.mode` property instead, which allows distinguishing between controller firmware, end device CLI firmware and the bootloader.
 - The `allowBootloaderOnly` driver option has been removed. To achieve the same behavior as before, set the new `bootloaderMode` option to `allow`.
 
+## Moved OTW firmware update functionality from `Controller` to `Driver` class
+
+To properly support OTW firmware updates in the different firmware types, the OTW firmware update functionality has been moved from the `Controller` class to the `Driver` class.
+
+- The `Controller.firmwareUpdateOTW` method has been removed. Use the new `Driver.firmwareUpdateOTW` method instead. The function signature is unchanged.
+- The `Controller.isFirmwareUpdateInProgress` method has been removed. Use the new `Driver.isOTWFirmwareUpdateInProgress` method instead. The function signature is unchanged.
+- The following types have been renamed:
+  - `ControllerFirmwareUpdateProgress` to `OTWFirmwareUpdateProgress`
+  - `ControllerFirmwareUpdateStatus` to `OTWFirmwareUpdateStatus`
+  - `ControllerFirmwareUpdateResult` to `OTWFirmwareUpdateResult`
+
+  ...and are no longer exported from the `zwave-js/Controller` or `zwave-js/Controller/safe` entry points. Import them from `zwave-js` or `zwave-js/Driver` instead.
+- The `"firmware update progress"` and `"firmware update finished"` events have been moved from the `Controller` class to the `Driver` class. Their signature is unchanged, except for the renamed types, see above.
+- In bootloader mode, the `driver.controller` property is now `undefined`, since it is no longer necessary to access the OTW firmware update functionality.
+
 ## Deprecated `isDocker` method has been removed
 
 Use the [`is-docker` package](https://github.com/sindresorhus/is-docker) instead or copy its code into your project.
