+++ packages/zwave-js/src/Controller_safe.ts
@@ -9,10 +9,7 @@ export type { ZWaveLibraryTypes } from "@zwave-js/core/safe";
 export type { ControllerStatistics } from "./lib/controller/ControllerStatistics.js";
 export { ZWaveFeature } from "./lib/controller/Features.js";
 export * from "./lib/controller/Inclusion.js";
-export { ControllerFirmwareUpdateStatus } from "./lib/controller/_Types.js";
 export type {
-	ControllerFirmwareUpdateProgress,
-	ControllerFirmwareUpdateResult,
 	FirmwareUpdateDeviceID,
 	GetFirmwareUpdatesOptions,
 	RebuildRoutesOptions,
