+++ packages/zwave-js/src/lib/controller/_Types.ts
@@ -63,30 +63,3 @@ export interface GetFirmwareUpdatesOptions {
 	 */
 	rfRegion?: RFRegion;
 }
-
-export interface ControllerFirmwareUpdateProgress {
-	/** How many fragments of the firmware update have been transmitted. Together with `totalFragments` this can be used to display progress. */
-	sentFragments: number;
-	/** How many fragments the firmware update consists of. */
-	totalFragments: number;
-	/** The total progress of the firmware update in %, rounded to two digits. */
-	progress: number;
-}
-
-export enum ControllerFirmwareUpdateStatus {
-	// An expected response was not received from the controller in time
-	Error_Timeout = 0,
-	/** The maximum number of retry attempts for a firmware fragments were reached */
-	Error_RetryLimitReached,
-	/** The update was aborted by the bootloader */
-	Error_Aborted,
-	/** This controller does not support firmware updates */
-	Error_NotSupported,
-
-	OK = 0xff,
-}
-
-export interface ControllerFirmwareUpdateResult {
-	success: boolean;
-	status: ControllerFirmwareUpdateStatus;
-}
